requirements.txt
aws-cdk-lib==2.83.1
aws-cdk.aws_lambda_python_alpha==2.83.1a0
constructs>=10.0.0,<11.0.0